Adding Asset class and modifiying Search, FileResult classes
